Allow page heading and <title> to differ
An upcoming commit introduces a page with quite a long page heading
which happens to include the name of the app in it. The resulting page title, with its
automatically suffixed app name, doesn't make for great reading!

This change allows us control over both independently, if and when we
want to use it.