Move non-legacy JS module into a separate sub-directory
I've used require_tree so that we won't need to change
app/assets/javascripts/application.js every time we add/remove a module.
This seems to be consistent with what other GOV.UK apps are doing.

I'm planning to add more non-legacy JS and this change will make that
easier.