Scope for never signed in Users that have expired
We're planning to start deleting unused user accounts. But "unused" is
hard to define. And it's unclear what potential impacts this might have
on the publishing apps that those users have accessed.

So, here's a definition of unused user accounts that definitely haven't
accessed any publishing apps! Users that were invited more than 90 days
ago but have never signed in.

We've chosen 90 days because

1. 90's round-ish number, as numbers-of-days go
2. it's comfortably longer than the 2 weeks that invitations expire after
3. it sounds like a generous enough amount of time for a new user and
   their inviter can go back and forth, attempting to get signed in,
   resending the invitation, etc

Since it appears that invitations can be resent, this query is using
`invitation_sent_at` (instead of, e.g. `created_at`) to give all
invitations the same grace period.