Rake task to delete expired never-signed-in Users
That is, web users (ie. not API users) who were invited 90+ days ago and
have never signed in.

Once this is deployed, we'll:

  1. trigger the initial run manually
  2. add a cron job to govuk-helm-charts to run this regularly

Unless we decide not to create the EventLogs, I'm not sure how this can
be made more efficient. That initial run took ~3 minutes to complete on
my machine (against a dump of the integration database).