User#application_permissions dependent: :destroy
Without this config, a destroyed User record will leave its associated
UserApplicationPermission records intact. Since they'll be completely
useless at that point, we don't want that to happen.

It appears that we've never deleted Users from the Signon database
before now, but we're planning to start doing so shortly -- in the very
specific scenario where a web user was invited to Signon 90+ days ago
but has never signed in.

Users do have a few other associated records that I've left unchanged,
mostly because none should exist yet for a user who's never signed in:

  1. BatchInvitations -- these are invitations created by the user, so
     can't belong to a user who's never signed in
  2. BulkGrantPermissionSets -- also created by the user
  3. EventLogs where user is the "initiator" -- I think these are also
     created by the user in the same sense as the two above (i.e. they
     record an event performed by one user against the account of
     another)
  4. Doorkeeper::AccessGrant -- is managed by the Doorkeeper gem
  5. Doorkeeper::AccessToken -- also managed by the Doorkeeper gem

Note: not all of the above are documented with a `has_many` in the User
model.

If we do end up deleting users in future who have previously been active
users, it's not obvious that we'd want to delete all of the above in
those cases either. For example, EventLogs are used as audit log
entries, so it seems like a bad idea to delete them along with the user
who initiated them.