Avoid queueing job if Splunk-related functionality disabled
Trello: https://trello.com/c/q1B1HKBy

Currently the Splunk-related functionality is effectively disabled,
because neither of the relevant env vars are defined in any environment.
See this Trello card [1] for more details.

Prior to this commit, we were always queueing up a Sidekiq job whenever
EventLog.record_event was called even though, as soon as that
SplunkLogStreamingJob was executed, the first thing we did was give up
if both the relevant env vars were not set.

Now we only queue up the job in the first place if the relevant env vars
are defined. I've left the original check in EventLog#send_to_splunk in
place, because there's a small chance the state of the env vars could
change between the job being queued and the job being executed.

[1]: https://trello.com/c/Zp23Kd6c