Make UserLockingTest more robust
I'm about to make some changes to EventLog.record_event and I don't want
those changes to affect this test.

Previously we were asserting that 8 failed sign-in attempts resulted in
10 enqueued jobs with a comment explaining that 9 of these jobs were the
result of calls to EventLog.record_event.

I've changed the test so that (a) it uses the value of
User.maximum_attempts defined in the Devise configuration rather than
the magic number 8 (which was actually incorrect; it should've been 6!);
and (b) it ensures a single email (as opposed to any old job) is
enqueued when the maximum number of attempts is reached.

I did attempt to add a more specific assertion using
ActionMailer::TestHelper#assert_enqueued_email_with, but I couldn't get
it to work.