Allow table helper #cell to take a block
Instead of a text argument, similar to how a lot of view helpers work.

But I'm going to avoid offering any flexibility in the order of
arguments (that's a common view helper thing when there's a choice
between text argument or block, right?). We'll just ignore the value
supplied for that first argument if there is a block