Move Created-at from API Users index to edit page
I spoke to two of the more frequent users of the API Users index and
edit pages about this. They said it's not normally useful, but it was
useful during replatforming because there were multiple API Users per
app (in that case, there was one for EKS and one for EC2, one created
hours or days ago and one created a few years ago).

I'm confident we can imagine other ways to solve that problem (using the
other columns) if it did happen to come up again. At a push, I think
we'd be better off just moving this column back into the table
temporarily than leaving it taking up space in the meantime.

I've moved the information onto the API User edit page instead of just
dropping it entirely purely because it is genuine information about the
API User. I can't say for sure that's it'll ever be useful there without
more research, though.