Copy Table component from gem
And make some small changes to turn it into an app-based component.

We're trying to make our tables more usable on small screens.

Since this goal doesn't appear to contradict the existing Table
component's goals, we'd like to merge those improvements into the gem.
But first of all, we need to develop and test the changes, so we're
planning to:
  1. ~~duplicate the gem's version of the component~~
  2. make the changes locally
  3. test them in our app, probably in production
  4. open a pull request to merge the changes into the gem
  5. switch the app back to using the gem's version

It's unclear what the cleanest strategy is to achieve this copying and
modifying. For instance, I've included CSS and JavaScript here although
we might have got away with just depending on the gem's versions? And
a hundred other dilemmas.

Changes I've already made to the copied files (to make the app not crash
and to turn this into an app-based component):
  - change helper name and namespacing, and make class a module -- the
    result is an unpleasant compromise between the gem's approach and
    how I'd expect the app to do this if starting from scratch
  - remove the `add_gem_component_stylesheet` line from the partial and
    don't replace it with `add_app_component_stylesheet` -- I'm not sure
    what the app one is meant to do, because it didn't appear to do what
    adding `@import "components/*";` to application.scss did
  - rename 'gem-c-table' and related CSS classes and JavaScript
    selectors to 'app-c-table', etc
  - remove inline warning comment from helper -- although part of me
    wants to add warning comments to all of these files that I've copied
  - add 1 Rubocop and 1 yarn stylelint ignore comment -- it's existing
    code and we want to keep it as close as possible to the original