Set API Users table to `vertical_on_small_screen`
Using the app's new version of the Table component, this lays the
table's contents out vertically on screens tablet-sized and below.

In the related controller test, I've switched to using a regexp instead
of a string to match the contents of the table cells. The implementation
of `vertical_on_small_screen` embeds some markup in each table cell,
which would break these strict matchers.