Move UserFilterHelper#assignable_user_roles -> UsersHelper
This method is nothing to do with the filtering that the rest of
this helper is concerned with. This method is only used in the
app/views/users/_form_fields.html.erb partial template which allows
setting or updating a user's role; whereas all the others are used in
the filtering of users in the users index page.