Remove filter functionality from users index page
Trello: https://trello.com/c/2gvEJKw9

This filtering functionality doesn't comform to the Design Sytem so
we're going to have to re-implement it so that it does. As a first step
I'm removing the old implementation which should hopefully make the
addition of the new functionality easier to follow.

I've inlined UserFilterHelper#user_role_text to a hard-coded string,
because it no longer varies according to the filtered role.