Introduce UsersFilter on users index page
Trello: https://trello.com/c/2gvEJKw9

This is a first step towards implementing the sidebar filtering in the
new proposed design. It's loosely based on Whitehall's
Admin::EditionsFilter [1]. I haven't actually implemented any filtering
in this commit - I've just moved the ordering and pagination into the
new UsersFilter in preparation for implementing filtering in the
UsersFilter#users method based on the options set from the params.

[1]: https://github.com/alphagov/whitehall/blob/8ddae783224de0c4889dbea0e3d049649287ee75/app/models/admin/edition_filter.rb