Improve test for exporting CSV of users
Previously the magic number, 3, in the assertion related to (a) the
admin user created in the setup; (b) the user created in the test; and
(c) the header row.

Now we parse the CSV so we don't have to consider the header row and we
use User.count to make it clear that we're including all users, i.e. in
this case both the admin user and the normal user.