Add name/email filtering to users page
Trello: https://trello.com/c/2gvEJKw9

This adds a search input field to the filter sidebar. When the user
enters a value in that input field and clicks either the integrated
search button or the form "Update results" button, the list of users is
restricted to only those which have a name or email partially matching
the search input value.

I've tried to give the new User scopes more intention-revealing names
than the previous implementation. And I've tried to make the new
implementation of the scopes more idiomatic Rails. Unlike the previous
implementation of the scopes, I'm not stripping the supplied string,
because I think this it's more appropriate for the caller to handle
this, i.e. UsersFilter#users in this case.

I considered wrapping the supplied string in a call to
ActiveRecord::Sanitization::ClassMethods#sanitize_sql_like to escape
underscores & percent signs. However, in the end I decided against doing
this because it would be a change from the previous behaviour and in any
case I can imagine it might be useful to use these wildcard characters
in searches.

I've used the slightly confusing "filter" param name, because that's
what was used in the previous implementation and I didn't want to
unnecessarily break any bookmarked URLs.