Add a "Clear all filters" link
Trello: https://trello.com/c/2gvEJKw9

I've wrapped this in a button group with the "Update results" button so
that the link appears alongside the button in wider screen widths, but
below the button in narrower screen widths.

I've added the govuk-link--no-visited-state CSS class to the link,
because that's what we've done elsewhere - it's not obvious it's useful
to highlight that the user has clicked on this link previously.