Copy option-select component from finder-frontend
The copy was made from the finder-frontend repo at this commit [1].

We're planning to use this component on the users index page for the
filter functionality in the proposed design.

I haven't copied any of the specs, because I would've had to translate
the RSpec spec into Minitest and I wasn't sure the Jasmine test would
just work. Since the plan is to extract this component into the
govuk_publishing_components gem, it didn't seem worth the effort.

[1]: https://github.com/alphagov/finder-frontend/tree/38758e110bf8397c64b2143771794318af9d25b5