Add role filtering to users page
Trello: https://trello.com/c/2gvEJKw9

This adds an instance of the new option-select component to the filter
sidebar. When the user selects one or more or the checkboxes and clicks
the form "Update results" button, the list of users is restricted to
only those which have one of the selected roles.

This represents a change from the orginal filtering implementation which
only allowed the user to filter by a single role. I have confirmed [1]
that this is an intentional change and not just an accidental
implication of the proposed design.

I've used `roles` as the parameter name to make it clear that more than
one role can be selected and to distinguish it from the previous
implementation which used `role`. We might want to consider redirecting
URLs using `role` to use `roles` if we want bookmarked URLs to continue
to work as expected.

[1]: https://trello.com/c/2gvEJKw9/207-migrate-users-page-to-use-design-system#comment-64e5f080394409b567e5bcb4