Add organisation filtering to users page
Trello: https://trello.com/c/2gvEJKw9

This adds an instance of the new option-select component to the filter
sidebar. When the user selects one or more or the checkboxes and clicks
the form "Update results" button, the list of users is restricted to
only those which belong to one of the selected organisations.

This represents a change from the orginal filtering implementation which
only allowed the user to filter by a single organisation. I have
confirmed [1] that this is an intentional change and not just an
accidental implication of the proposed design.

I've used `organisations` as the parameter name to make it clear that
more than one organisation can be selected and to distinguish it from
the previous implementation which used `organisation`. We might want to
consider redirecting URLs using `organisation` to use `organisations` if
we want bookmarked URLs to continue to work as expected.

The option-select component is only displayed if the current user is
allowed to manage more than one organisation. I've made this more
explicit than the original implementation by using
User#manageable_organisations in
UsersFilter#organisation_option_select_options.

If the current user is allowed to manage exactly one organisation (i.e.
they are an organisation admin) then the filtered users heading is
qualified with the relevant organisation, e.g. "399 users in Government
Digital Service". This is because the organisation option select will
not be displayed in this scenario so we need to make it clear which
organisation the users in the table belong to.

[1]: https://trello.com/c/2gvEJKw9/207-migrate-users-page-to-use-design-system#comment-64e5f080394409b567e5bcb4