Add permission filtering to users page
Trello: https://trello.com/c/2gvEJKw9

This adds an instance of the new option-select component to the filter
sidebar. When the user selects one or more or the checkboxes and clicks
the form "Update results" button, the list of users is restricted to
only those which have one of the selected permissions.

This represents a change from the orginal filtering implementation which
only allowed the user to filter by a single permission. I have confirmed
[1] that this is an intentional change and not just an accidental
implication of the proposed design.

I've used `permissions` as the parameter name to make it clear that more
than one permission can be selected and to distinguish it from the
previous implementation which used `permission`. We might want to
consider redirecting URLs using `permission` to use `permissions` if we
want bookmarked URLs to continue to work as expected.

I've used arel_table to generate the column name in the existing
User.with_partially_matching_name & User.with_partially_matching_email
scopes so that they work in conjunction with the new
User.with_permission scope, i.e. to avoid ambiguity in column names.

I've tried to make the implementation of
UsersHelper#permission_option_select_options a bit more idioomatic than
the original version in UserFilterHelper#user_filter_list_items:

* The new implementation doesn't specify an order, because we can rely
  on the default scope order of Doorkeeper::Application &
  SupportedPermission.
* I didn't feel as if plucking values made the query significantly more
  efficient and it cluttered up the code.
* Using a map loop within a flat_map loop allowed me to use instances of
  Doorkeeper::Application & SupportedPermission which made the code a
  bit simpler / more readable.

[1]: https://trello.com/c/2gvEJKw9/207-migrate-users-page-to-use-design-system#comment-64e5f080394409b567e5bcb4