Add status filtering to users page
Trello: https://trello.com/c/2gvEJKw9

This adds an instance of the new option-select component to the filter
sidebar. When the user selects one or more or the checkboxes and clicks
the form "Update results" button, the list of users is restricted to
only those which have one of the selected statuses.

This represents a change from the orginal filtering implementation which
only allowed the user to filter by a single status. I have confirmed [1]
that this is an intentional change and not just an accidental
implication of the proposed design.

I've used `statuses` as the parameter name to make it clear that more
than one status can be selected and to distinguish it from the previous
implementation which used `status`. We might want to consider
redirecting URLs using `status` to use `statuses` if we want bookmarked
URLs to continue to work as expected.

Unlike for the previous implementation of User.with_status which used a
case statement to return the relevant un-named scope, for
User.with_statuses I've implemented a bunch of named scopes which make
it easier to combine them programmatically using
ActiveRecord::QueryMethods#or. This makes the code easier to test and I
think it makes it easier to read. I'm hoping we might be able to re-use
the named scopes elsewhere in the code. I've also introduced some new
user factories for the various user "states".

Note that the new User.locked scope perpetuates the issue where we don't
consider Devise's time-based unlock strategy, i.e. where
User#access_locked? may return false when the user is still included in
the User.locked scope because User#locked_at is not nil. This problem is
captured in this Trello card [2].

[1]: https://trello.com/c/2gvEJKw9/207-migrate-users-page-to-use-design-system#comment-64e5f080394409b567e5bcb4
[2]: https://trello.com/c/88jMyKuf