Add 2SV status filtering to users page
Trello: https://trello.com/c/2gvEJKw9

This adds an instance of the new option-select component to the filter
sidebar. When the user selects one or more or the checkboxes and clicks
the form "Update results" button, the list of users is restricted to
only those which have one of the selected 2SV statuses.

This represents a change from the orginal filtering implementation which
only allowed the user to filter by a single 2SV status. I have confirmed [1]
that this is an intentional change and not just an accidental
implication of the proposed design.

I've used `two_step_statuses` as the parameter name to make it clear
that more than one 2SV status can be selected and to distinguish it from
the previous implementation which used `two_step_status`. We might want
to consider redirecting URLs using `two_step_status` to use
`two_step_statuses` if we want bookmarked URLs to continue to work as
expected.

Unlike for the previous implementation of User.with_2sv_enabled which
used a case statement to return the relevant un-named scope, for
User.with_2sv_statuses I've implemented a bunch of named scopes which
make it easier to combine them programmatically using
ActiveRecord::QueryMethods#or. This makes the code easier to test and I
think it makes it easier to read. I'm hoping we might be able to re-use
the named scopes elsewhere in the code.

[1]: https://trello.com/c/2gvEJKw9/207-migrate-users-page-to-use-design-system#comment-64e5f080394409b567e5bcb4