Add horizontal lines to separate filters on users page
Trello: https://trello.com/c/2gvEJKw9

As per the propposed design.

I couldn't see a way to style the search component directly without
targetting *all* instances of the component, so I've wrapped it in a div
element with a CSS class which adds the bottom border.

I've used the same CSS class to add a bottom border to the button group
which was already in a containing div.