Open option-select components if any options selected
This makes the option-select components on the users page "open" if any
of its options are selected.

The option-select for statuses will also "open" if none of the
option-select components have any options selected - this is to make it
clearer to users what's hidden away in the "closed".

I think this will also make things less confusing when we add JS to
submit the form when any of the checkboxes changes state.