Put selected options first when there are a lot of options
There are a lot of options in the option-select component for
organisations & permissions on the users index page. While it's easy
enough to find a specific option using the search box (displayed because
show_filter is set to true), it's not so easy to find the options that
have already been selected unless you can remember the option label.

Putting the selected options above all the unselected options should
make things much easier.