Redirect legacy users filters on users index page
The filters on the users index page now allow multiple options to be
selected and the param keys have changed from singular to plural. Also
the values for the two-step status filter have changed to match the
relevant User scopes.

This change redirects requests to the users index page that use the
legacy filter params. This is in case anyone has bookmarked some of
these URLs. Note that the redirection won't handle all URLs containing a
mixture of legacy and non-legacy params, but this seems like a real edge
case that's not worth worrying about.

It might be worth adding a flash message to instruct users to update
their bookmark so that we can safely remove this functionality in the
not too distant future, but I haven't done that here.