Automatically submit users filter form
This introduces a new AutoSubmitForm JS module which is activated for
the form on the users index page. This JS automatically submits the form
whenever the form's change event fires, i.e. whenever one of the filter
checkboxes is checked/unchecked.

The "auto-submit-ignore" data attribute should be a comma-separated list
of input names for which to ignore the form's change event. We use this
for the "option-select-filter" text inputs which are dynamically added
to the option-select component when `show_filter` is set to `true`.
The values of those text inputs are not used on the server-side and so
we don't want to auto-submit the form when they change.