Move 3 buttons on users index page to top-right
As per the proposed design [1].

I'm not particularly happy with the changes to the admin_layout. I think
the real problem is that we've got too much of the page in the layout
and we can probably come up with a better way to avoid duplication.
However, I haven't attempted to address that here, because it would be a
significant chunk of work and this PR is already big enough!

Since the container isn't using a flex layout as suggested by the
prototype, I had to use `float: right` to position the button group to
the right of the page on wider screens.

I'm not 100% sure whether I've used the correct naming conventions in
the naming of the "users-index-button-group" CSS class.

[1]: https://trello.com/c/bXsVUK6c