Remove <abbr> permissions list on app names
This tooltip was on our list of things that don't seem like they'd
comply with the design system.

I've now spoken to a few recent and frequent users of the API Users
index page and of the ones who knew this feature existed, none have used
it, so I'm happy to remove it.

The tests that used this feature are now using the API User edit page to
get that information.