Increase max-width of govuk-width-container for all pages
As agreed in the collaboration session on Wed 06 Sep [1], we're
increasing the value of `$govuk-page-width` from the default of 960px to
1140px which is the value used in Whitehall [2]. We're making this
change in order to give more room to the table on the new version of the
users index page which now has a filter sidebar occupying valuable
screen wdith. This will reduce the chances of the table overflowing the
right-hand side of the page in screens wider than the "tablet"
breakpoint, i.e. when the table is still in "horizontal" mode.

[1]: https://trello.com/c/2gvEJKw9/207-migrate-users-page-to-use-design-system#comment-64f87c053ccc070a1fae1821
[2]: https://github.com/alphagov/whitehall/blob/b355ef2caf4b414487d3687495d133b66f7f7846/app/assets/stylesheets/application.scss#L1