Use custom breakpoint for table on users page
As agreed in the collaboration session on Wed 06 Sep [1], we're
increasing the value of this breakpoint from the standard "tablet"
breakpoint of 641px to a custom breakpoint of 940px. We chose this value
so that the table changes into "vertical" mode before the table content
overflows the right-hand side of the container and before a scrollbar
appears at the bottom of the page.

This change also affects the API users index page, but seems to work OK
there too.

[1]: https://trello.com/c/2gvEJKw9/207-migrate-users-page-to-use-design-system#comment-64f87c053ccc070a1fae1821