Allow GOV.UK Admins to give themselves access to apps
This adds a "Grant access" button next to each app in the list of "Apps
you don't have access to" on /account/applications. This page is
currently only available to GOV.UK Admins and this functionality
replicates what they're already able to do on their /users/<id>/edit page.

I experimented with using the button component for the "Grant access"
button but couldn't work out how to pass the visually hidden span into
it.

I considered using a more generic `PermissionsController` (instead of
`SigninPermissionsController`) given that the signin permission is
modelled the same as all other permissions, and that we're going to
allow users to manage other permissions in future. I decided against it
for now because the signin permission feels conceptually different to
the other permissions. We can revisit this decision in future if it
turns out there's a better approach.