Add test coverage for helper method
I'm planning on introducing a new "status" to BatchInvitations and
I believe this helper method is where that concept currently lives.

I've avoided the temptation to test every possible scenario
exhaustively. Before I looked this closely, though, I'd imagined that
this status and BatchInvitation#outcome mapped 1:1, and that
BatchInvitation#outcome and BatchInvitationUser#outcome were related.
I hope that these tests are clear enough to demonstrate that neither of
those assumptions are true.