Remove unwanted whitespace from existing Organisation records
There is an `Organisation` record in our production data that has
leading and/or trailing whitespace in its name column. This migration
removes that whitespace. Additional whitespace means records don't
sort alphabetically in an expected way.

At the time of writing there's only a single record with this issue in
Integration and none in Production (although there's no guarentees
others won't have been added by the time this migration runs). Because
there's such a small number of records it seems safe to load and
resave these records, triggering the hook added in the previous
commit.