Strip unwanted whitespace from Organisation name
Organisations with unnecessary whitespace don't sort in an expected
way. Before fixing exisiting records with this problem, this commit
prevents any new occurrences.

Rails 7.1 introduces ActiveRecord::Base::normalizes[1] which could be
used to simplify this code. That hasn't yet been released and it seems
a bit unconventional within GDS to depend on edge rails, but when 7.1
is out, we should be able to adopt that pattern.

[1] https://github.com/rails/rails/pull/43945