Remove unreachable code in batch_invitations/new.html.erb
The `BatchInvitationPolicy#assign_organisation_from_csv?` method is an
alias for `BatchInvitationPolicy#new` which means that only people who
can view this page can assign organisations from the CSV file. So the
second branch (where `assign_organisation_from_csv` is `false`) is not
reachable in this view template and can therefore be removed.