Remove table of recent batches from batch_invitations/new.html.erb
The new (design-system) designs for this page don't include the table
of recent batches, so in preparation for migrating this page to the
design system, I'm removing them.

With the removal of the table the `batch_invite_status_link` helper
method is no longer used so can also be removed.