Migrate BatchInvitationsController#new to design system
This commit switches the first "page" of the batch creation of users
workflow to the design system.

All of the actions in BatchInvitationsController now use the
`admin_layout` so I've removed the `only` array rather than add `new`
to it.

The `form_for` helper needs `multipart: true` for file uploads to
work. This confused me for a while - I think rails default form
builder knows to set `multipart: true` implicitly if you use the
`file_field` helper, but as we've removed that in order to use the
`file_upload` component we now have to set it explicitly.

I've inlined the relevant `policy_scope` line from
`UserHelpers#organisation_options` to build the list of
`Organisations` in the select. This helper method is still used
elsewhere though, so I can't remove it yet.