Make Active the first User status
Active was last on the list of user status filter options in the filters
section of the Users#index page, but it seems like it would be a more
common choice than the others, so this commit moves it to the top of the
list of options.

I haven't found any other user-facing changes that the order of this
array would effect.