Remove BulkGrantPermissionSet feature
We don't believe that this feature is used anymore. Or that it should be
in future (bulk-granting access to abitrary publishing apps for all
users at the click of a button could be considered dangerous).

It was originally introduced as part of the implementation of [RFC
078](https://github.com/alphagov/govuk-rfcs/blob/8cbb2a0de86de02f54ae37b245e79b46ad62cb6a/rfc-078-re-architect-signin-permissions-in-signon.md)
("Signon MUST allow a super admin to grant permissions to all users in
one go."). Signon's authorisation data model hasn't evolved much since
that time, so we're not confident that that need for bulk granting has
gone away entirely (especially for "signin" permission). But, the same
results can be achieved without needing to expose and maintain this
functionality in the web interface, so we've concluded that we're better
off without it.