Inline shared/user_permissions template
We want to make fairly extensive changes to how the permissions are
selected for this page of the batch user creation process and convert
it to the design system. Inlining the shared form will let me make
some changes to the logic first and then hopefully make it easier to
see the design-system changes that follow.