Inline `applications_and_permissions`
The second block variable yielded by this helper method is not used,
instead we only need the applications.

This list of applications was previously obtained from
`UserPermissionsControllerMethods#applications_and_permissions` which in
turn called `UserPermissionsControllerMethods#visible_applications` with
the `user_object` which in this view is always `User.new`. Because
`User.new.api_user? == false` the implementation of
`UserPermissionsControllerMethods#visible_applications` fell through
to return `policy_scope(:user_permission_manageable_application)`. So
we can inline this in the view instead of the more convoluted chain of
methods.