Remove unnecessary argument to `options_for_select`
In this `user_object` is always `User.new` and
`User.new.permission_ids_for(application)` will always return `[]`
since a new User does not have any associated
`application_permissions`. This means that no applications need to be
pre-selected in this call to `options_for_select` and we can rely on
the default value for the second argument (which is `nil`).