Replace permission select box with checkboxes
In the new designs for this page we want to have a list of checkboxes
rather than the current `chosen`-powered select box. As a step towards
that this commit replaces the select box with checkboxes.

I haven't made any attempt to style these new checkboxes since they
will shortly be replaced by design-system ones.

With this commit it is less obvious that we need to have a distinction
between the sign-on permission and the others, but I'm going to live
with the rather ugly chained `reject` method here and tidy that up next.