Migrate BatchInvitationPermissionsController#new to design system
This commit migrates the BatchInvitationPermissionsController#new page
to the design system. Hopefully with the preparatory commits this diff
isn't too hard to follow.

I've had to use the "raw" HTML version of the accordian component[1]
rather than the one in the publishing component library[2] since the
latter doesn't support passing a block of arbitrary content into each
section. I think it's a small enough amount of HTML to take this route
rather than try to extend the component.

The breadcrumb trail could be improved - ideally it would link to the
previous "step" in this flow which would involve editing the CSV file
and organisation - but we don't have an edit route for this yet and
I'm concerned already about the size of this change. I think it's one
we could revisit in a subsequent enhancement.

[1] https://design-system.service.gov.uk/components/accordion/
[2] https://components.publishing.service.gov.uk/component-guide/accordion