Allow GOV.UK Admins to remove their access to app
Users are required to confirm the removal of access to avoid the problem
that exists in the current `/users/<id>/edit` page where Publishing
Managers can accidentally remove their access to an application but then
have to open a support request to get that access back again. NOTE.  The
functionality in this PR is currently restricted to GOV.UK Admins but
will be made available to Publishing Managers soon.

I experimented with using the button component for the "Remove access"
button but couldn't work out how to pass the visually hidden span into
it.