Move "Change your email or password" to /account
We're making a new Account page. The new page will link to a handful of
sub-pages that each handle a specific account-related task.

This existing page will be used for changing your email address or
password, so I've moved it into the /account namespace to hopefully make
the user journey feel more cohesive.