Allow GOV.UK Admins to view their app permissions
This adds the `/account/applications/<id>/permissions` page which lists
all app permissions and whether the current user has that permission or
not.

This new page is currently only accessible to GOV.UK Admins that have
the signin permission to the app.

The list of permissions has "signin" first, followed by:

- other permissions the user has been granted; ordered alphabetically
- other permissions the user has not been granted; ordered alphabetically