Use redirect_back_or_to in user_not_authorized
If the user attempts to do something they're not authorized to do then
we should first try to take them back to the page they were on, and only
redirect to the root path as a fallback. Ideally users wouldn't be able
to use the UI to navigate to actions they're not authorized to execute
but this small change will make the experience slightly better if they
are able to.